Resolved circular dependencies
Closes #1190
