+++ packages/core/src/service/__tests__/find-address-with-pk.test.ts
@@ -6,9 +6,10 @@ import {
   SIGN_ALGO_NUM_ECDSA_P256,
 } from '@onflow/flow-wallet-shared/constant/algo-constants';
 
-import { findAddressWithPK, findAddressWithSeed } from '../findAddressWithPK';
-import * as findAddressWithPubKeyModule from '../findAddressWithPubKey';
-import * as publicPrivateKeyModule from '../publicPrivateKey';
+import { findAddressWithPK, findAddressWithSeed } from '@/service/account-management';
+
+import * as publicPrivateKeyModule from '../../utils/modules/publicPrivateKey';
+import { getOrCheckAccountsByPublicKeyTuple } from '../account-management';
 
 // Mock the dependent modules
 vi.mock('../publicPrivateKey');
@@ -72,25 +73,20 @@ describe('findAddressWithPK module', () => {
         },
       ];
 
-      vi.mocked(
-        findAddressWithPubKeyModule.getOrCheckAccountsByPublicKeyTuple
-      ).mockResolvedValueOnce(mockAccounts);
+      vi.mocked(getOrCheckAccountsByPublicKeyTuple).mockResolvedValueOnce(mockAccounts);
 
       const result = await findAddressWithPK(mockPK, mockAddress);
 
       expect(publicPrivateKeyModule.pk2PubKeyTuple).toHaveBeenCalledWith(mockPK);
-      expect(findAddressWithPubKeyModule.getOrCheckAccountsByPublicKeyTuple).toHaveBeenCalledWith(
-        mockPubKeyTuple,
-        mockAddress
-      );
+      expect(getOrCheckAccountsByPublicKeyTuple).toHaveBeenCalledWith(mockPubKeyTuple, mockAddress);
       expect(result).toEqual(mockAccounts);
     });
 
     it('should throw error when no accounts are found', async () => {
       vi.mocked(publicPrivateKeyModule.pk2PubKeyTuple).mockResolvedValueOnce(mockPubKeyTuple);
-      vi.mocked(
-        findAddressWithPubKeyModule.getOrCheckAccountsByPublicKeyTuple
-      ).mockRejectedValueOnce(new Error('No accounts found with the given public key'));
+      vi.mocked(getOrCheckAccountsByPublicKeyTuple).mockRejectedValueOnce(
+        new Error('No accounts found with the given public key')
+      );
 
       await expect(findAddressWithPK(mockPK, mockAddress)).rejects.toThrow(
         'No accounts found with the given public key'
@@ -117,9 +113,7 @@ describe('findAddressWithPK module', () => {
         },
       ];
 
-      vi.mocked(
-        findAddressWithPubKeyModule.getOrCheckAccountsByPublicKeyTuple
-      ).mockResolvedValueOnce(mockAccounts);
+      vi.mocked(getOrCheckAccountsByPublicKeyTuple).mockResolvedValueOnce(mockAccounts);
 
       const result = await findAddressWithSeed(mockSeed, mockAddress);
 
@@ -129,10 +123,7 @@ describe('findAddressWithPK module', () => {
         ''
       );
       expect(publicPrivateKeyModule.seed2PublicPrivateKeyTemp).not.toHaveBeenCalled();
-      expect(findAddressWithPubKeyModule.getOrCheckAccountsByPublicKeyTuple).toHaveBeenCalledWith(
-        mockPubKeyTuple,
-        mockAddress
-      );
+      expect(getOrCheckAccountsByPublicKeyTuple).toHaveBeenCalledWith(mockPubKeyTuple, mockAddress);
       expect(result).toEqual(mockAccounts);
     });
 
@@ -154,17 +145,12 @@ describe('findAddressWithPK module', () => {
         },
       ];
 
-      vi.mocked(
-        findAddressWithPubKeyModule.getOrCheckAccountsByPublicKeyTuple
-      ).mockResolvedValueOnce(mockAccounts);
+      vi.mocked(getOrCheckAccountsByPublicKeyTuple).mockResolvedValueOnce(mockAccounts);
 
       const result = await findAddressWithSeed(mockSeed, mockAddress);
 
       expect(publicPrivateKeyModule.seed2PublicPrivateKey).not.toHaveBeenCalled();
-      expect(findAddressWithPubKeyModule.getOrCheckAccountsByPublicKeyTuple).toHaveBeenCalledWith(
-        mockPubKeyTuple,
-        mockAddress
-      );
+      expect(getOrCheckAccountsByPublicKeyTuple).toHaveBeenCalledWith(mockPubKeyTuple, mockAddress);
       expect(result).toEqual(mockAccounts);
     });
 
@@ -172,9 +158,9 @@ describe('findAddressWithPK module', () => {
       vi.mocked(publicPrivateKeyModule.seed2PublicPrivateKey).mockResolvedValueOnce(
         mockPubKeyTuple
       );
-      vi.mocked(
-        findAddressWithPubKeyModule.getOrCheckAccountsByPublicKeyTuple
-      ).mockRejectedValueOnce(new Error('No accounts found with the given public key'));
+      vi.mocked(getOrCheckAccountsByPublicKeyTuple).mockRejectedValueOnce(
+        new Error('No accounts found with the given public key')
+      );
 
       await expect(findAddressWithSeed(mockSeed, mockAddress)).rejects.toThrow(
         'No accounts found with the given public key'
