+++ packages/core/src/service/keyring/index.ts
@@ -41,9 +41,12 @@ import {
   consoleWarn,
 } from '@onflow/flow-wallet-shared/utils/console-log';
 
+import { getAccountsByPublicKeyTuple } from '@/utils/key-indexer';
+
+import { HDKeyring, type HDKeyringData, type HDKeyringType } from './hdKeyring';
+import { type SimpleKeyPairType, SimpleKeyring, type SimpleKeyringData } from './simpleKeyring';
 import { normalizeAddress } from '../../utils';
 import { defaultAccountKey, pubKeyAccountToAccountKey } from '../../utils/account-key';
-import { getAccountsByPublicKeyTuple } from '../../utils/modules/findAddressWithPubKey';
 import {
   formPubKeyTuple,
   getPublicKeyFromPrivateKey,
@@ -51,8 +54,6 @@ import {
   seedWithPathAndPhrase2PublicPrivateKey,
 } from '../../utils/modules/publicPrivateKey';
 import preference from '../preference';
-import { HDKeyring, type HDKeyringData, type HDKeyringType } from './hdKeyring';
-import { type SimpleKeyPairType, SimpleKeyring, type SimpleKeyringData } from './simpleKeyring';
 
 export const KEYRING_SDK_TYPES = {
   SimpleKeyring,
