+++ packages/core/tsconfig.json
@@ -8,6 +8,8 @@
     "declarationMap": true,
     "sourceMap": true,
     "strict": true,
+    "checkJs": true,
+    "noImplicitAny": false,
     "esModuleInterop": true,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
@@ -23,7 +25,7 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["src/**/*", "tsup.config.ts", "vitest.config.ts", "vitest.init.ts"],
+  "include": ["src/**/*", "tsup.config.ts", "vitest.config.ts", "vitest.init.ts", "build/**/*"],
   "exclude": ["node_modules", "dist"],
 
   "references": [
@@ -32,6 +34,9 @@
     },
     {
       "path": "../data-model"
+    },
+    {
+      "path": "../extension-shared"
     }
   ]
 }
