+++ packages/data-model/src/batch-refresh-integration.test.ts
@@ -1,11 +1,9 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { getCachedData } from '@onflow/flow-wallet-data-model/cache-data-access';
-import {
-  accountBalanceKey,
-  accountBalanceRefreshRegex,
-} from '@onflow/flow-wallet-data-model/cache-data-keys';
 import storage from '@onflow/flow-wallet-extension-shared/storage';
+
+import { getCachedData } from './cache-data-access';
+import { accountBalanceKey, accountBalanceRefreshRegex } from './cache-data-keys';
 
 // Mock the storage module
 vi.mock('@onflow/flow-wallet-extension-shared/storage', () => ({
@@ -17,7 +15,7 @@ vi.mock('@onflow/flow-wallet-extension-shared/storage', () => ({
 }));
 
 // Mock getCachedData
-vi.mock('@onflow/flow-wallet-data-model/cache-data-access', () => ({
+vi.mock('./cache-data-access', () => ({
   getCachedData: vi.fn(),
 }));
