+++ packages/data-model/src/data-cache.ts
@@ -1,6 +1,7 @@
-import { type CacheDataItem } from '@onflow/flow-wallet-data-model/data-cache-types';
 import storage from '@onflow/flow-wallet-extension-shared/storage';
 import { consoleError } from '@onflow/flow-wallet-shared/utils/console-log';
+
+import { type CacheDataItem } from './data-cache-types';
 
 /**
  * Get valid data from session storage
